Added the execution interface. Agents can now dispatch actions to the executor, which will monitor for completion. Next steps are to add the ability to modify action execution, and to process action feedback from the dispatcher side

+++ install/ma_interfaces/include/ma_interfaces/ma_interfaces/msg/detail/bid__struct.h
@@ -27,8 +27,8 @@ typedef struct ma_interfaces__msg__Bid
 {
   rosidl_runtime_c__String agent_id;
   rosidl_runtime_c__String auction_id;
-  int64_t st;
-  int64_t et;
+  float st;
+  float et;
   int64_t value;
 } ma_interfaces__msg__Bid;
